reduce compiler warnings in KOKKOS version of USER-DPD code

+++ src/KOKKOS/pair_sw_kokkos.cpp
@@ -130,8 +130,8 @@ void PairSWKokkos<DeviceType>::compute(int eflag_in, int vflag_in)
 
   int max_neighs = d_neighbors.extent(1);
 
-  if ((d_neighbors_short.extent(1) != max_neighs) ||
-     (d_neighbors_short.extent(0) != ignum)) {
+  if ((int) (d_neighbors_short.extent(1) != max_neighs) ||
+      (int) (d_neighbors_short.extent(0) != ignum)) {
     d_neighbors_short = Kokkos::View<int**,DeviceType>("SW::neighbors_short",ignum,max_neighs);
   }
   if (d_numneigh_short.extent(0)!=ignum)
